feat: Game의 기능 Controller와 합치기
